Refactor module names for PEP 8 compliance
