+++ tests/test_hdf_writer.py
@@ -1,9 +1,9 @@
 import h5py
 from math import sqrt
 from pytest import approx
-from geometry_constructor.Writers import HdfWriter
-from geometry_constructor.DataModel import Detector, PixelGrid, Corner, CountDirection, Vector
-from geometry_constructor.QmlModel import InstrumentModel
+from geometry_constructor.writers import HdfWriter
+from geometry_constructor.datamodel import Detector, PixelGrid, Corner, CountDirection, Vector
+from geometry_constructor.qmlmodel import InstrumentModel
 
 
 def assess_unit_length_3d_vector(vector, original):
